requirements.txt
pymupdf
pillow
groq
chromadb
sentence-transformers
python-dotenv
pydantic
rich
pathlib